refactor: move metrics under experimental namespace
Relocate the metrics package to openretailscience.experimental.metrics to
signal its experimental status. Update internal imports, tests, docs, and
mkdocs navigation to reflect the new path.
