--- docs/api/metrics/distribution.md
@@ -1,5 +1,0 @@
-# Distribution Metrics
-
-::: openretailscience.metrics.distribution.acv
-
-::: openretailscience.metrics.distribution.pct_of_stores
