+++ openretailscience/experimental/__init__.py
@@ -1,0 +1,4 @@
+"""Experimental modules.
+
+APIs in this package are experimental and may change without notice.
+"""
