+++ tests/experimental/metrics/distribution/test_pct_of_stores.py
@@ -1,11 +1,11 @@
-"""Tests for openretailscience.metrics.distribution.pct_of_stores."""
+"""Tests for openretailscience.experimental.metrics.distribution.pct_of_stores."""
 
 import ibis
 import pandas as pd
 import pytest
 from pandas.testing import assert_frame_equal
 
-from openretailscience.metrics.distribution.pct_of_stores import PctOfStores
+from openretailscience.experimental.metrics.distribution.pct_of_stores import PctOfStores
 from openretailscience.options import ColumnHelper, get_option
 
 cols = ColumnHelper()
